requirements.txt
tensorflow-gpu==1.2.1
scipy==1.2.1
pillow==6.0.0
scikit_image==0.15.0